requirements.txt
Flask==3.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-Session==0.5.0
google-generativeai==0.3.2
gunicorn==21.2.0